HUD::LateUpdate 호출 대신, 변동이 있을때만 상수버퍼 갱신하도록 변경

+++ David Engine Project/Engine Source/daStructs.h
@@ -12,8 +12,8 @@ namespace da::structs
 	};
 	struct sDashCount
 	{
-		float MaxCount;
-		float CurCount;
+		int MaxCount;
+		int CurCount;
 		sDashCount()
 			: MaxCount(0), CurCount(0)
 		{}
